Add item to FAQ about cb
Includes Plan 9 source code and a html rendered man page. If there was a
unix or linux version the code can no longer be found.

Some README.md files refer to this tool. When an author's remarks it
will not be removed or updated but if in the judges' remarks it might be
moved to a historical note subsection. This will happen in time with
additional passes of the README.md files - for most it should be the
final pass but the years 2011+ will need a pass before that.