Add indent.c to .gitignore
This is because of the make rule 'indent' which runs 'indent' on the
files, saving to 'indent.c'.

If any entry ever has a file called indent.c something will have to
change obviously but for now this will prevent a cluttered git status if
someone runs make indent (like I did yesterday or the day before).