Updates and fixes to 2013/dlowe
Added fun.c and make make all compile it: this is what the judges
suggested one try compiling with different compilers and checking the
results. I added how to specify the compiler.

Added slen.sh script, corrected for shellcheck, which runs the command
the author suggested (and which the judges added to their remarks), to
make it simpler for viewers of the entry.

Added INABIAF section to README.md: if no args are specified it will
very likely crash or do something funny. A similar such thing is in
bugs.md as well.

Updated .gitignore.

Format fix README.md.