Format fix 2013/endoh1
This also involved undoing the rot13 spoiler which uncovered some typos
as well. I did not rot13 it again as Landon has specified that unless an
entry is to be used to decipher text in the documentation (an example
being my Enigma machine, 2020/ferguson2) then it should not be
enciphered as the point of remarks in the contest has changed in some
ways.