Fix 2013/hou + format 2013/hou/README.md
The way the Makefile was constructed meant that the program just
decompressed the real program which is only the first part of the
program. As the judges noted it could be said to break rule 2 (in a
way?) because the code actually is a decompressor of the real code. So
in order to get the entry to do what it's supposed to do the Makefile
rule was modified to be:

    ${PROG}: ${PROG}.c
	    ${CC} ${CFLAGS} $< -o $@ ${LDFLAGS}
	    ./${PROG} | ${CC} ${CFLAGS} -x c - -o $@ ${LDFLAGS}

which first compiles the program to hou and then runs hou, piping the
output to the compiler to compile it into hou again. This way when one
runs hou (after running make) the actual program is run, saving
viewers/users from the extra step (and potential confusion). The
README.md file also shows how to just generate the decompressed source
code and notes the author's remarks about a.c.gz and a.c (what was
included in what was a.c.gz) so as to clarify that the file is not
missing but rather that it's not included and does not need to be
included either.

Since the real code uses math.h the LDFLAGS were modified to include -lm.

The .gitignore file was updated to have the PPM files that are generated
by the program.

Added an INAFIAB section to the README.md and the bugs.md file: it does
not terminate on its own and since the author mentioned it wrt a rule it
felt like it should be put in such a section.

Format fixed README.md.

With this commit only a final pass should be necessary to say this entry
has been finished though it's entirely possible I discover another
problem I guess.