Fix fix on fix of 1990/tbr
When I added the 'gets()' and 'exit()' macros to make the code look more
like the original I accidentally left out a space between 'gets(x)' and
'fgets(...' so that it looked like:

    #define gets(x)fgets((x),512,stdin)

which I believe is incorrect. It now looks instead like:

    #define gets(x) fgets((x),512,stdin)

The compact version that the authors provided was compile fixed quite
some time back but I never added the exit macro to match the original
more closely.

Also since the compact version was meant to be identical in function of
the entry and since that entry was changed to use fgets() (by the macro
gets()) I decided it would be better to make it use fgets() also, also
through redefining gets(). To make the line length closer in match to
the original compact version I changed the code to be two lines. This
wasn't necessary to do, as I recall, in the entry itself though it's
possible it was and I just managed to get the ASCII art the same or
almost the same.

The thanks-for-fixes.md file was updated for this change but
unfortunately I had made some changes yesterday that I forgot about
(just rewording) in another entry so that will also be in this commit.