Minor fix at top of thanks file
The file previously noted several thousand changes that were mentioned
above but this was mentioned at the top of the file so there were no
several thousand changes. There is something like that in the news but
that won't stay there - in time it'll be rotated out - and the context
is missing anyway. Thus I have removed the part about the changes being
above and slightly modified the rest to be clearer of what was done.