Swap alt/main in 1985/lycklama
This is a rare case where it's better to first try the alternate code as
it moves too quickly to see in modern systems. Thus the build/try
section refers to alt and it explains how to reconfigure the time for
usleep(3). The original version is described in the Original code
section. It is encouraged to try this after trying the alternate code.

I also updated the thanks file to note that part of the fix to this
entry was the inclusion of unistd.h.

With this commit the README.md has also had a final pass of format/typo
checks and fixes though it might be that the way the alt/original
build/try section is written might be standardised in which case this
file would have to be modified again.