Make my 1984/decot fix look more like original
I was able to bring back some code and only had to comment out one part
and make a minor change to another another. Originally the code had:

    extern int floor;

but this is not possible in modern systems so it was changed to:

    extern double floor(double);

The

    double (x1, y1) b;

could be brought back, changed from:

    int b,

(the comma is not a typo there).

Now it looks rather more like the original.

The thanks-for-fixes.md file was updated here, including a diff to try
(to see what had to be done to fix the entry for modern systems) but I'm
afraid it has a minor change in another entry, something I accidentally
left out of 1984/mullender.