Add to FAQ about seeing fix changes
It seems to me that it is a good idea to show people how to see what
changed in order to get an entry to work in modern systems. I gave an
example where it's harder to tell because it's a very long line and an
example where it's easier to tell as it's not long lines but rather
shorter lines. Well it's easier to see on a line-by-line basis ..
whether it's easy to see what actually changed is probably another
matter entirely (especially since it's easier to see when you have
compiler errors in front of you) :-)