Fix 1986/pawka/Makefile
The flag -Wno-strict-prototypes was put in the CDEFINE variable,
suggesting that it might need a -D flag when it of course does not. This
is not strictly an error but it's also not strictly correct so I
strictly decided to make it correct.

At the same time I added to CSILENCE -Wno-implicit-int, a warning that
has shown up in macOS recently, very possibly with macOS Sonoma, and so
might have to be added to many Makefiles.