Make 1987/heckbert look more like original
As index(3) in some systems requires strings.h and since it's identical
to strchr and since the judges said in System V machines (not that we
should have to worry about those these days) they had to define index to
strchr I have done this in the Makefile.

Also the #define _ define was restored even though it's not used in
order to make it look closer to the original entry. This has to be done
in other files as well, as I think on it.

Format and typo checked README.md.