Make 1987/westley more symmetrical after fix
With the compilation fix of the code some symmetry was unfortunately
destroyed. To try and keep it as close as possible to the original I had
just commented out the code. But looking at it again today it occurred
to me that I can make it look more symmetrical by removing the commented
out code and then formatting the code a bit differently (another cast
was removed to do this). It is not any less authentic but it now has
more symmetry which was an important part of the entry's layout.