Update 1988/litmaath
Format and typo check the README.md file.

Improved the order of the alt code (displayed by the judges) and what
followed it (that might seem like it's related to the alt code when
although maybe true it still is about the entry itself).

Updated thanks file to refer to the alt code and removed myself from the
README.md (this before the far superior idea of a single file outside of
the entries existed and because some entries were missed, this one
included).