Format/typo fix 1988/reddy + restore winner name
By some crazy chance the winner's name was not in the README.md file:
just a corporation (which has been removed along with the address). I
was able to locate it in the author/ subdirectory in the JSON file.

Format and typo checked the README.md and also added some things to try
to explore the entry more.