Add 1988/westley/demo.sh + format/typo check
The reason for the demo is because when one looks at the code and then
sees the output at the same time it's far more beautiful and magical.

The README.md has been typo and format checked.

With this commit, the README.md files for the 1988 entries have been
checked and besides some things brought up on GitHub they should be in
good order (for now?).