Fix 1987/lievaart - check for invalid input
It was documented that the program detected invalid input but it
actually (now - it used to work) printed in an infinite loop 'You:'. Now
doing invalid input forfeits the game which is better than flooding the
screen. It is not yet clear if it should act like you put in 99 (see the
README.md for details) so this is in bugs.md just in case. If it is
supposed to just skip a move then this can be fixed but otherwise it can
be removed from bugs.md without any change in the code. This fix applies
to both versions, the one with the board display and the original entry
(with contest restrictions on size of code).