Fix 1989/jar.2
There were some things unclear in the README.md file that have now been
corrected.

Clarified in the README.md that running make clobber all will run the
program itself.

The make alt target now runs the alt code just like the entry itself.
This is fine because it is a simulation (to avoid annoyance of writing
binary data to /dev/tty) and since one may still run the program or
script directly it's fine.

The alternate code section in the README.md was cleaned up and formatted
as well.