Minor update to 1989/jar.1/README.md
This is done so that I can provide another thing I forgot to mention in
the previous log. I fixed the scripts to not have the ill-famed useless
use of cat. Why is it ill-famed?

Because there's no such thing as a useless cat, that's why! In fact the
fact they exist proves that the Earth is NOT flat: if it was flat they
would have knocked everything off it by now! :-) Besides, they offer all
us ailurophiles hours and hours and years and years of joy and fun and
companionship. And contrary to popular belief, they most certainly do
have a personality and they can be very sociable.