Correct thanks in 1990/scjones
.. i.e. remove it from the README.md file, this time of Yusuke Endoh
(not his fault at all but rather a problem that occurred before the
superior idea of a single file was thought up).

I had already format fixed this file and the only other thing I did was
remove the address (but this is not strictly necessary as the author
information will eventually be generated from the JSON files).