Improve my 1991/dds alt code (required for clang)
Based on a slight modification to the Makefile the program will now
compile to the same file as if dds.c was being used (a.out).

I updated the README.md with some changes in formatting to do with the
alt code but the file has not been fully corrected or even checked.