Updates to 1984/decot wrt alt/orig code
Update formatting of the alt code (heading levels) and added subsections
for how to build and run (rather than having it all together).

Updated the historical remarks as well. It's important in this case to
refer to decot.orig.c and not decot.alt.c because the orig code is the
one that was submitted whereas the alt code is the modification the
judges made.