Slightly improve (I hope :-) ) bugs.md statuses
What this means is I changed the heading level from 2 to 3 of the STATUS
lines (under the entries) so that it shows that the status is for the
respective entries.

This was thought of when I was working on moving the bugs/(mis)features
of the entries of 1980s to bugs.md (as in making the README.md files
refer to the bugs.md but not include the details, just the status) and
noticed that it just didn't look quite right. Hopefully this improves it
just a bit.