Rename how-to-bugfix.md to how-to-help.md
This is because it doesn't really say how to fix a bug (how could it
when the bugs can be all sorts of different things) but rather a
more general how to help with bugs and other things. It's not just bugs
but other issues with entries or even the website. Thus how-to-help.md
is a better name.