Improve test script (now scripts) for 1995/garry
Added alt test script to use alt code (though in this case it has no
real use other than to be complete). It makes sure to compile the alt
code before trying to use it, exiting if it fails.

The original test script now also makes sure to compile the program
before use, exiting if it fails.

I also fixed the thanks file with this: I didn't fix the main entry;
that already compiled. I had fixed the alt code which did not compile
(partly because I agreed with the judges that it looked like a rat
dumping core).

Why an alt test script instead of just adding an environmental variable
like I did with others? Because it felt like it'd be easier to type. It
might be that I end up doing this with the scripts I added a way to
specify the program, when I come across it, if it seems worth it (which
for some it might very well be worth it).