Add missing ':' to Alternate code sections
This was done as:

    sgit -e 's/\(# Alternate code\)$/\1:/g' '*README.md'

The heading levels and other things to do with alternate code that are
being discussed have not been addressed here. That will come later. This
is, as noted, just adding the missing ':' where it was missing.

The command does NOT touch it when it is not prefixed with '# ' because
there are cases where normal lines end with 'Alternate code' (these will
be turned into links later on).