Add alt code 1989/westley
Unfortunately the fix to let the main program compile with clang even
breaks some versions generated (see compile.sh) in syntax error so even
gcc cannot compile those. Therefore it's now alt code. The alt version
will compile with clang and the main feature of the program will work
but the program cannot generate every version of the program.

The bugs.md and thanks-for-fixes.md have been updated along with the
README.md.

In the README.md there is a different heading level for the alt code
sections to compare to another as we try and figure out what heading
levels to use.

I changed the delay to 0 in the compile.sh script but one can set it to
another value if they like.

This entry will have to be worked on more at another time.