Change '#### Try' to '### Alternate try'
This does not complete the alt code related headers but I noticed this
and it was easy to do quickly. Unless I run across another when I am
working on the README.md files of the entries I will probably have to do
the rest manually as I come across them as most likely the headers are
missing.

This was done as

    sgit -e 's/^#### Try/### Alternate try/g' '*README.md'