Make thanks/complicated/etc. stand out in thanks
As the GitHub bold is rather close in colour to regular text (at least
in my admittedly poor vision) I have made the thanks, the complicated
and technically challenging stand out more by making them all upper
case.

I also made the wording (of thanks to me) closer to that of Yusuke but
changing in his 'VERY' to 'EXTREMELY': better to have it consistent and
some were quite challenging.

For entries listed under my name at the top I did not add the word
technically as I think that that's implied.