Typo fixes in 199[01]/dds
In the README.md file of 1990/dds there was a typo fixed.

In the thanks file for 1990/dds I clarified something and in the case of
1991/dds there was a typo that was confusing: '-Wno-error' should have
been '-Werror'; the fix enabled in the encrypted string the former but
the problem was some compilers in some platforms (clang, macOS) default
to -Werror which prevented the entry from compiling code as the code
generated had warnings.