Rename demo.sh scripts to try.sh
This is because they are under the try section. I'm aware that there's a
script called try.me.sh but that's not been decided on a name change. It
probably should be changed to try.sh to be consistent and probably will
be but it's not been yet.

The manifest files in tmp/ have been updated as well, including the
macOS Numbers file.