Remove 1986/hague alt code, make main entry
I fixed it so that the form where it would segfault (after printing the
correct output) no longer occurs.

The thanks file has had a note added towards the top about changing to
fgets(), noting that although more is mentioned in the FAQ, it can be a
real problem with some systems. The best example of it being beneficial
is 1990/tbr. Naturally I also updated the thanks file for 1986/hague
too, updating the details.